Week 1 - MP & MRP
Thoughts:

I know that my code is very basic, but coding with e.g. "tuples" is very unfamiliar to me, so I've done what I thought was reasonable.
I do not really understand the discrepancy between using r(s,s') or R(s) - I know that we have defined r(s,s') as the expected value of the reward in the current time, given that we know the current time - but wouldn't that just be deterministic then?